SRS has crashed before first keyframe arriving on rtc2rtmp.

+++ trunk/src/app/srs_app_rtc_source.hpp
@@ -305,7 +305,7 @@ private:
     const static uint16_t s_cache_size = 512;
     RtcPacketCache cache_video_pkts_[s_cache_size];
     uint16_t header_sn_;
-    uint16_t lost_sn_;
+    int32_t lost_sn_;
     int64_t rtp_key_frame_ts_;
 public:
     SrsRtmpFromRtcBridger(SrsLiveSource *src);
